+++ fabric/src/main/java/net/mehvahdjukaar/moonlight/api/client/util/fabric/RenderUtilImpl.java
@@ -22,7 +22,7 @@ import net.minecraft.world.level.block.state.BlockState;
 
 public class RenderUtilImpl {
 
-    public static void renderBlock(long seed, PoseStack poseStack, MultiBufferSource buffer, BlockState state, Level level, BlockPos pos, BlockRenderDispatcher blockRenderer) {
+    public static void renderBlock(BakedModel model, long seed, PoseStack poseStack, MultiBufferSource buffer, BlockState state, Level level, BlockPos pos, BlockRenderDispatcher blockRenderer) {
         blockRenderer.getModelRenderer().tesselateBlock(level, blockRenderer.getBlockModel(state), state, pos, poseStack, buffer.getBuffer(ItemBlockRenderTypes.getMovingBlockRenderType(state)),
                 false, RandomSource.create(), seed, OverlayTexture.NO_OVERLAY);
     }